+++ source/B2B.Transactions.CimMessageAdapter.Tests/B2B.Transactions.CimMessageAdapter.Tests.csproj
@@ -32,28 +32,28 @@
       <None Update="RequestChangeOfSupplier_postman.xml">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
-      <None Update="Messages\InvalidRequestChangeOfSupplier.xml">
+      <None Update="InvalidRequestChangeOfSupplier.xml">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
-      <None Update="Messages\RequestChangeOfSupplierWithDuplicateTransactionIds.xml">
+      <None Update="RequestChangeOfSupplierWithDuplicateTransactionIds.xml">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
-      <None Update="Messages\urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd">
+      <None Update="urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
-      <None Update="Messages\urn-entsoe-eu-local-extension-types.xsd">
+      <None Update="urn-entsoe-eu-local-extension-types.xsd">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
-      <None Update="Messages\urn-entsoe-eu-wgedi-codelists.xsd">
+      <None Update="urn-entsoe-eu-wgedi-codelists.xsd">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
-      <None Update="Messages\ValidRequestChangeOfSupplier.xml">
+      <None Update="ValidRequestChangeOfSupplier.xml">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\B2B.CimMessageAdapter\B2B.CimMessageAdapter.csproj" />
+      <ProjectReference Include="..\B2B.Transactions.CimMessageAdapter\B2B.Transactions.CimMessageAdapter.csproj" />
     </ItemGroup>
 
 </Project>
