+++ source/B2B.Transactions.CimMessageAdapter.Tests/RequestChangeOfSupplierWithDuplicateTransactionIds.xml
