+++ source/B2B.Transactions.CimMessageAdapter.Tests/ValidRequestChangeOfSupplier.xml
