+++ source/B2B.Transactions.CimMessageAdapter.Tests/urn-entsoe-eu-local-extension-types.xsd
