+++ source/B2B.Transactions.CimMessageAdapter.Tests/urn-entsoe-eu-wgedi-codelists.xsd
