+++ source/B2B.Transactions.CimMessageAdapter/B2B.Transactions.CimMessageAdapter.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <TargetFramework>net5.0</TargetFramework>
+        <RootNamespace>B2B.CimMessageAdapter</RootNamespace>
     </PropertyGroup>
 
 </Project>
