+++ source/B2B.Transactions.CimMessageAdapter/IMarketActivityRecordForwarder.cs
