+++ source/B2B.Transactions.CimMessageAdapter/IMessageIds.cs
