+++ source/B2B.Transactions.CimMessageAdapter/ITransactionIds.cs
