+++ source/B2B.Transactions.CimMessageAdapter/MarketActivityRecord.cs
