+++ source/B2B.Transactions.CimMessageAdapter/Result.cs
