+++ source/B2B.Transactions.CimMessageAdapter/ValidationError.cs
